Sort scripts into directories & update namespaces
