+++ Client/vr-client/Assets/NanoverIMD/Subtle Game/MudraGlove/mudra-2fingers.png
