+++ Client/vr-client/Assets/NanoverIMD/Subtle Game/MudraGlove/mudra-3fingers.png
