+++ Client/vr-client/Assets/NanoverIMD/Subtle Game/MudraGlove/mudra-hands-2fingers.mat.meta
