+++ Client/vr-client/Assets/NanoverIMD/Subtle Game/MudraGlove/mudra-hands-3fingers.mat.meta
