+++ Client/vr-client/Assets/NanoverIMD/Subtle Game/MudraGlove/support/hand-mudra-model.blend.meta
