+++ Client/vr-client/Assets/NanoverIMD/Subtle Game/MudraGlove/support/mudra-2fingers-nowrist.png.meta
