+++ Client/vr-client/Assets/NanoverIMD/Subtle Game/MudraGlove/support/mudra-og-texture-3fingers.png.meta
