+++ Client/vr-client/Assets/NanoverIMD/Subtle Game/MudraGlove/support/mudra-texture-workbench.xcf.meta
