+++ Client/vr-client/Assets/NanoverIMD/Subtle Game/Multiplayer/NanoverImdAvatarManager.cs
@@ -4,10 +4,9 @@ using Nanover.Core.Math;
 using Nanover.Frontend.Utility;
 using Nanover.Frontend.XR;
 using Nanover.Grpc.Multiplayer;
+using NanoverIMD.Subtle_Game.Multiplayer;
 using UnityEngine;
 using UnityEngine.XR;
-
-using NanoverImd.UI;
 
 namespace NanoverImd
 {
