--- Client/vr-client/Assets/NanoverIMD/UI.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4095f6ac0bbf71c449355c560398b8ad
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
